Refactor: RP-100 TestController 생성
테스트/개발용 API들을 묶은 컨트롤러임.
